// src/app/models/document.model.ts
export class DocumentModel {
    created_at: Date = null;
    document_id: string = '';
    document_picture: string = '';
    expired_date: Date = null;
    is_document_expired: boolean = false;
    is_expired_date: boolean = false;
    is_unique_code: boolean = false;
    is_uploaded: number = 0;
    name: string = '';
    option: number = 0;
    _id: string = "";
    user_id: string = '';
    token:string= '';
    __v: number = 0;
    updated_at: Date = null;
    unique_code: string = '';
}